Add express server for serving build; Create layout component and fix routing

+++ src/app/core/stores/tags.store.ts
@@ -1,13 +1,13 @@
-import {computed, inject, Injectable, signal} from "@angular/core";
+import {computed, effect, inject, Injectable, signal} from "@angular/core";
 import {TagsService} from "../services/tags.service";
 import {ITag} from "../models";
 import {IOptionItem} from "../../shared/models/option-item";
 import {map, Subject, switchMap, tap} from "rxjs";
+import {AuthStore} from "./auth.store";
 
-@Injectable({
-    providedIn: 'root'
-})
+@Injectable({providedIn: 'root'})
 export class TagsStore {
+    private _authStore = inject(AuthStore);
     private _tags = signal<ITag[]>([])
 
     tags = computed(() => this._tags());
@@ -24,20 +24,25 @@ export class TagsStore {
     delete$ = new Subject<number>();
 
     constructor() {
-        this.fetch$.subscribe(res => this._tags.set(res));
-        this.create$.pipe(
-            switchMap(payload => this.service.create(payload)),
-            tap(res => this._tags.update(prev => [...prev, res]))
-        ).subscribe();
+        effect(() => {
+            console.log("tagstore")
+            if (this._authStore.isLoggedIn()) {
+                this.fetch$.subscribe(res => this._tags.set(res));
+                this.create$.pipe(
+                    switchMap(payload => this.service.create(payload)),
+                    tap(res => this._tags.update(prev => [...prev, res]))
+                ).subscribe();
 
-        this.update$.pipe(
-            switchMap(payload => this.service.update(payload)),
-            tap(res => this._tags.update(prev => prev.map(s => s.id === res.id ? res : s)))
-        ).subscribe();
+                this.update$.pipe(
+                    switchMap(payload => this.service.update(payload)),
+                    tap(res => this._tags.update(prev => prev.map(s => s.id === res.id ? res : s)))
+                ).subscribe();
 
-        this.delete$.pipe(
-            switchMap(id => this.service.delete(id).pipe(map(() => id))),
-            tap((id) => this._tags.update(prev => prev.filter(s => s.id !== id)))
-        ).subscribe()
+                this.delete$.pipe(
+                    switchMap(id => this.service.delete(id).pipe(map(() => id))),
+                    tap((id) => this._tags.update(prev => prev.filter(s => s.id !== id)))
+                ).subscribe()
+            }
+        });
     }
 }
