+++ src/app/items/store/item.store.ts
@@ -1,12 +1,12 @@
-import {computed, inject, Injectable, signal} from "@angular/core";
+import {computed, effect, inject, Injectable, signal} from "@angular/core";
 import {map, Subject, switchMap, tap} from "rxjs";
 import {IItem} from "../models/item.model";
 import {ItemService} from "../service/item.service";
+import {AuthStore} from "../../core/stores/auth.store";
 
-@Injectable({
-    providedIn: "root"
-})
+@Injectable({providedIn: "root"})
 export class ItemStore {
+    private _authStore = inject(AuthStore);
     private _items = signal<IItem[]>([])
 
     items = computed(() => this._items());
@@ -18,20 +18,24 @@ export class ItemStore {
     delete$ = new Subject<number>();
 
     constructor() {
-        this.fetch$.subscribe(res => this._items.set(res));
-        this.create$.pipe(
-            switchMap(payload => this.service.create(payload)),
-            tap(res => this._items.update(prev => [...prev, res]))
-        ).subscribe();
+        effect(() => {
+            if (this._authStore.isLoggedIn()) {
+                this.fetch$.subscribe(res => this._items.set(res));
+                this.create$.pipe(
+                    switchMap(payload => this.service.create(payload)),
+                    tap(res => this._items.update(prev => [...prev, res]))
+                ).subscribe();
 
-        this.update$.pipe(
-            switchMap(payload => this.service.update(payload)),
-            tap(res => this._items.update(prev => prev.map(s => s.id === res.id ? res : s)))
-        ).subscribe();
+                this.update$.pipe(
+                    switchMap(payload => this.service.update(payload)),
+                    tap(res => this._items.update(prev => prev.map(s => s.id === res.id ? res : s)))
+                ).subscribe();
 
-        this.delete$.pipe(
-            switchMap(id => this.service.delete(id).pipe(map(() => id))),
-            tap((id) => this._items.update(prev => prev.filter(s => s.id !== id)))
-        ).subscribe()
+                this.delete$.pipe(
+                    switchMap(id => this.service.delete(id).pipe(map(() => id))),
+                    tap((id) => this._items.update(prev => prev.filter(s => s.id !== id)))
+                ).subscribe()
+            }
+        });
     }
 }
