+++ src/app/savings/savings.store.ts
@@ -1,12 +1,12 @@
-import {computed, inject, Injectable, signal} from "@angular/core";
+import {computed, effect, inject, Injectable, signal} from "@angular/core";
 import {ISaving} from "./models";
 import {SavingsService} from "./savings.service";
 import {map, Subject, switchMap, tap} from "rxjs";
+import {AuthStore} from "../core/stores/auth.store";
 
-@Injectable({
-    providedIn: 'root'
-})
+@Injectable({providedIn: "root"})
 export class SavingsStore {
+    private _authStore = inject(AuthStore);
     private service = inject(SavingsService)
     private _savings = signal<ISaving[]>([])
 
@@ -18,20 +18,24 @@ export class SavingsStore {
     delete$ = new Subject<number>();
 
     constructor() {
-        this.fetch$.subscribe(res => this._savings.set(res));
-        this.create$.pipe(
-            switchMap(payload => this.service.create(payload)),
-            tap(res => this._savings.update(prev => [...prev, res]))
-        ).subscribe();
+        effect(() => {
+            if (this._authStore.isLoggedIn()) {
+                this.fetch$.subscribe(res => this._savings.set(res));
+                this.create$.pipe(
+                    switchMap(payload => this.service.create(payload)),
+                    tap(res => this._savings.update(prev => [...prev, res]))
+                ).subscribe();
 
-        this.update$.pipe(
-            switchMap(payload => this.service.update(payload)),
-            tap(res => this._savings.update(prev => prev.map(s => s.id === res.id ? res : s)))
-        ).subscribe();
+                this.update$.pipe(
+                    switchMap(payload => this.service.update(payload)),
+                    tap(res => this._savings.update(prev => prev.map(s => s.id === res.id ? res : s)))
+                ).subscribe();
 
-        this.delete$.pipe(
-            switchMap(id => this.service.delete(id).pipe(map(() => id))),
-            tap((id) => this._savings.update(prev => prev.filter(s => s.id !== id)))
-        ).subscribe()
+                this.delete$.pipe(
+                    switchMap(id => this.service.delete(id).pipe(map(() => id))),
+                    tap((id) => this._savings.update(prev => prev.filter(s => s.id !== id)))
+                ).subscribe()
+            }
+        });
     }
 }
